requirements.txt
numpy
opencv-python
mediapipe
pytesseract
transformers
torch
torchvision
sounddevice
speechrecognition
pyttsx3
face_recognition
ultralytics
python-dotenv
pyaudio
mediapipe==0.10.9
pytesseract==0.3.10
requests
geopy
timezonefinder
python-dotenv
pytz